refactor(demo): relocate bipartite matching and eulerian path demo wrappers to implementation files
